pyproject.toml
[project]
name = "pdf-summarise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.1.0",
    "pydantic>=2.11.9",
    "pypdf>=6.1.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
]
